requirements.txt
click==8.1.3
csscompressor==0.9.5
future==0.18.2
ghp-import==2.1.0
htmlmin==0.1.12
importlib-metadata==4.12.0
Jinja2==3.1.2
joblib==1.1.0
jsmin==3.0.1
livereload==2.6.3
lunr==0.6.2
Markdown==3.3.7
MarkupSafe==2.1.1
mergedeep==1.3.4
<<<<<<< temp
<<<<<<< temp2
mkdocs==1.4.2
mkdocs-material==9.0.15
mkdocs-material-extensions>=1.0.3
mkdocs-minify-plugin>=0.5.0
||||||| ancestor
mkdocs==1.3.1
mkdocs-material==8.4.0
mkdocs-material-extensions==1.0.3
mkdocs-minify-plugin==0.5.0
mkdocs-roamlinks-plugin==0.2.0
=======
mkdocs==1.4.2
mkdocs-material==9.0.15
mkdocs-material-extensions>=1.0.3
mkdocs-minify-plugin>=0.5.0
<<<<<<< temp2
mkdocs-roamlinks-plugin>=0.2.0
>>>>>>> Copied over changes from testing environment.
||||||| ancestor
mkdocs-roamlinks-plugin>=0.2.0
=======
>>>>>>> Updating repository from testing environment.
||||||| ancestor
mkdocs==1.3.1
mkdocs-material==8.4.0
mkdocs-material-extensions==1.0.3
mkdocs-minify-plugin==0.5.0
mkdocs-roamlinks-plugin==0.2.0
=======
mkdocs==1.4.2
mkdocs-material==9.0.15
mkdocs-material-extensions>=1.0.3
mkdocs-minify-plugin>=0.5.0
<<<<<<< temp
mkdocs-roamlinks-plugin>=0.2.0
>>>>>>> Copied over changes from testing environment.
||||||| ancestor
mkdocs-roamlinks-plugin>=0.2.0
=======
>>>>>>> Updating repository from testing environment.
nltk==3.7
packaging==21.3
Pygments>=2.13.0
pygtrie==2.5.0
pymdown-extensions>=9.5
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
pyyaml_env_tag==0.1
regex==2022.8.17
six==1.16.0
tornado==6.2
tqdm==4.64.0
watchdog==2.1.9
zipp==3.8.1
<<<<<<< temp
<<<<<<< temp2
wheel==0.38.4
python-frontmatter==1.0.0
mkdocs-ezlinks-plugin>=0.1.14
mkdocs-blogging-plugin>=2.2.2
#mkdocs-roamlinks-plugin>=0.2.0
git+https://github.com/peddamat/mkdocs-roamlinks-plugin@master
||||||| ancestor
=======
wheel==0.38.4
python-frontmatter==1.0.0
mkdocs-ezlinks-plugin>=0.1.14
mkdocs-blogging-plugin>=2.2.2
<<<<<<< temp2
>>>>>>> Copied over changes from testing environment.
||||||| ancestor
=======
#mkdocs-roamlinks-plugin>=0.2.0
git+https://github.com/peddamat/mkdocs-roamlinks-plugin@master
>>>>>>> Updating repository from testing environment.
||||||| ancestor
=======
wheel==0.38.4
python-frontmatter==1.0.0
mkdocs-ezlinks-plugin>=0.1.14
mkdocs-blogging-plugin>=2.2.2
<<<<<<< temp
>>>>>>> Copied over changes from testing environment.
||||||| ancestor
=======
#mkdocs-roamlinks-plugin>=0.2.0
git+https://github.com/peddamat/mkdocs-roamlinks-plugin@master
>>>>>>> Updating repository from testing environment.
